pyproject.toml
#############################
# fetcher – uv / hatch config #
#############################

[project]
name = "fetcher"
version = "0.1.9"
description = "Standalone release of the SciLLM/Sparta fetcher pipeline with Brave + SciLLM alternates."
readme = "README.md"
authors = [{name = "Graham Anderson", email = "graham@grahama.co"}]
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
  "aiohttp>=3.10",
  "beautifulsoup4>=4.12",
  "charset-normalizer>=3.3",
  "python-dotenv>=1.0",
  "requests>=2.32",
  "lxml>=5.3",
  "playwright>=1.48",
  "pymupdf>=1.24",
  "pytesseract>=0.3",
  "pillow>=10.4",
  "ftfy>=6.2",
  "readability-lxml>=0.8.1",
  "trafilatura>=2.0.0",
  "pdferli>=0.11",
]

[project.urls]
Homepage = "https://github.com/grahama1970/fetcher"
Documentation = "https://docs.litellm.ai/docs/fetcher"
Source = "https://github.com/grahama1970/fetcher"

[project.optional-dependencies]
rolling = [
  "spacy>=3.7,<4",
]
alternates = [
  "scillm @ file:///home/graham/workspace/experiments/litellm",
]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.23",
  "ruff>=0.6",
  "mypy>=1.9",
]
pdfcrack = [
  "pdferli>=0.1.0",
]

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fetcher"]

[dependency-groups]
telemetry = ["jsonlines>=4.0"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.mypy]
python_version = "3.12"
warn_unused_ignores = true

[tool.ruff]
line-length = 120
